Usability improvement: handle dtype conversion in data loader.
The model requires both the image and mask have single-precision
dtype. The way things are currently structured, we require users
to make this conversion on their own. IMO this is something that
should be handled internally for the sake of UX.

+++ deepcell_types/dataset.py
@@ -26,7 +26,10 @@ class PatchDataset(IterableDataset):
     ):
         super(PatchDataset, self).__init__(**kwargs)
 
-        self.mask = mask
+        # Model requires image and mask in single precision
+        raw = raw.astype(np.float32)
+        self.mask = mask.astype(np.float32)
+
         self.dct_config = dct_config
         self.max_channels = dct_config.MAX_NUM_CHANNELS
         self.paddings = -1.0
